# 资源/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: Hope and Sparks API
  version: 1.0.0
  description: 'Apifox 导入版接口规范。详细业务说明、数据库落库规则和枚举映射见根目录

    `Hope_and_Sparks_API接口文档.md`。

    '
servers:
- url: http://localhost:8080
  description: 本地开发环境
- url: /
  description: 相对路径
security:
- bearerAuth: []
...
paths:
...
  /api/v1/learning-plans/generate:
    post:
      tags:
      - 学习计划日历与拓扑
      summary: Strict 生成学习计划
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                courseId:
                  type: string
                  description: 课程 ID，对应 course.id
                goal:
                  type: string
                  description: 学习目标
                deadline:
                  type: string
                  example: '2026-06-30'
                  format: date
                  description: 截止日期
                difficulty:
                  type: string
                  enum:
                  - easy
                  - medium
                  - hard
                  - adaptive
                  description: 难度
                dailyMinutes:
                  type: integer
                  example: 60
                  description: 每日学习分钟数
                preferredTimeSlots:
                  type: array
                  items:
                    type: string
                  description: 偏好的学习时间段
              required:
              - courseId
              - goal
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `courseId` | 是 | string | 课程 ID，对应 course.id |

        | Body | `goal` | 是 | string | 学习目标 |

        | Body | `deadline` | 否 | string | 截止日期 |

        | Body | `difficulty` | 否 | string | 难度 |

        | Body | `dailyMinutes` | 否 | integer | 每日学习分钟数 |

        | Body | `preferredTimeSlots` | 否 | array<string> | 偏好的学习时间段 |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
  /api/v1/learning-plans/{planId}/adjust:
    put:
      tags:
      - 学习计划日历与拓扑
      summary: 调整学习计划
      parameters:
      - $ref: '#/components/parameters/planId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason:
                  type: string
                  description: 操作原因
                strategy:
                  type: string
                  enum:
                  - speed_up
                  - slow_down
                  - remove_redundant_review
                  - add_review
                  - manual
                  description: 计划调整策略
                taskIds:
                  type: array
                  items:
                    type: string
                  description: 学习任务 ID 列表，对应 study_task.id
              required:
              - reason
              - strategy
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `planId` | 是 | string | 学习计划 ID，对应 study_plan.id |

        | Body | `reason` | 是 | string | 操作原因 |

        | Body | `strategy` | 是 | string | 计划调整策略 |

        | Body | `taskIds` | 否 | array<string> | 学习任务 ID 列表，对应 study_task.id |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
  /api/v1/learning-plans/{planId}/topology:
    get:
      tags:
      - 学习计划日历与拓扑
      summary: 获取学习拓扑
      parameters:
      - $ref: '#/components/parameters/planId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `planId` | 是 | string | 学习计划 ID，对应 study_plan.id |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
  /api/v1/learning-plans/{planId}/topology/nodes/{nodeId}/resource-network:
    get:
      tags:
      - 学习计划日历与拓扑
      summary: 获取知识点资源网络图
      parameters:
      - $ref: '#/components/parameters/planId'
      - $ref: '#/components/parameters/nodeId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `planId` | 是 | string | 学习计划 ID，对应 study_plan.id |

        | Path | `nodeId` | 是 | string | 知识点节点 ID，对应 knowledge_node.id，兼容 knowledge_node.node_code |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
  /api/v1/learning-plans/{planId}/topology/nodes/{nodeId}/resources:
    get:
      tags:
      - 学习计划日历与拓扑
      summary: 获取知识点资源网络图列表视图
      parameters:
      - $ref: '#/components/parameters/planId'
      - $ref: '#/components/parameters/nodeId'
      - $ref: '#/components/parameters/type'
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          $ref: '#/components/responses/PageSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `planId` | 是 | string | 学习计划 ID，对应 study_plan.id |

        | Path | `nodeId` | 是 | string | 知识点节点 ID，对应 knowledge_node.id，兼容 knowledge_node.node_code |

        | Query | `type` | 否 | string | 类型筛选，具体枚举见接口文档数据字典 |

        | Query | `page` | 否 | integer | 页码，从 1 开始 |

        | Query | `pageSize` | 否 | integer | 每页数量 |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
  /api/v1/calendar/events:
    get:
      tags:
      - 学习计划日历与拓扑
      summary: 获取日历事件
      parameters:
      - name: month
        in: query
        schema:
          type: string
          example: 2026-05
      - name: view
        in: query
        schema:
          type: string
          enum:
          - month
          - week
          - day
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Query | `month` | 否 | string | 月份，格式 YYYY-MM |

        | Query | `view` | 否 | string | 日历视图类型：month/week/day |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
    post:
      tags:
      - 学习计划日历与拓扑
      summary: 新建日程
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                  description: 标题
                  example: 二叉搜索树章节课后练习
                startTime:
                  type: string
                  format: date-time
                  description: 开始时间
                endTime:
                  type: string
                  format: date-time
                  description: 结束时间
                type:
                  type: string
                  enum:
                  - learn
                  - review
                  - practice
                  - exam
                  - custom
                  description: type
                planId:
                  type: string
                  description: 学习计划 ID，对应 study_plan.id
                taskId:
                  type: string
                  description: 学习任务 ID，对应 study_task.id
                resourceId:
                  type: string
                  description: 学习资源 ID，对应 learning_resource.id
                remark:
                  type: string
                  description: 备注
              required:
              - title
              - startTime
              - endTime
              - type
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `title` | 是 | string | 标题 |

        | Body | `startTime` | 是 | string | 开始时间 |

        | Body | `endTime` | 是 | string | 结束时间 |

        | Body | `type` | 是 | string | type |

        | Body | `planId` | 否 | string | 学习计划 ID，对应 study_plan.id |

        | Body | `taskId` | 否 | string | 学习任务 ID，对应 study_task.id |

        | Body | `resourceId` | 否 | string | 学习资源 ID，对应 learning_resource.id |

        | Body | `remark` | 否 | string | 备注 |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
  /api/v1/calendar/events/{eventId}:
    put:
      tags:
      - 学习计划日历与拓扑
      summary: 更新日程
      parameters:
      - $ref: '#/components/parameters/eventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                  description: 标题
                  example: 二叉搜索树章节课后练习
                startTime:
                  type: string
                  format: date-time
                  description: 开始时间
                endTime:
                  type: string
                  format: date-time
                  description: 结束时间
                type:
                  type: string
                  enum:
                  - learn
                  - review
                  - practice
                  - exam
                  - custom
                  description: type
                status:
                  type: string
                  enum:
                  - pending
                  - in_progress
                  - completed
                  - overdue
                  description: status
                remark:
                  type: string
                  description: 备注
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `eventId` | 是 | string | 日历事件/学习任务 ID，对应 study_task.id |

        | Body | `title` | 否 | string | 标题 |

        | Body | `startTime` | 否 | string | 开始时间 |

        | Body | `endTime` | 否 | string | 结束时间 |

        | Body | `type` | 否 | string | type |

        | Body | `status` | 否 | string | status |

        | Body | `remark` | 否 | string | 备注 |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
    delete:
      tags:
      - 学习计划日历与拓扑
      summary: 删除日程
      parameters:
      - $ref: '#/components/parameters/eventId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `eventId` | 是 | string | 日历事件/学习任务 ID，对应 study_task.id |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
...
  /api/v1/skill-tree/nodes/{nodeId}/light-up:
    post:
      tags:
      - 技能树
      summary: 点亮技能节点
      parameters:
      - $ref: '#/components/parameters/nodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                evidenceType:
                  type: string
                  enum:
                  - resource
                  - exercise_set
                  - manual
                  description: 点亮依据类型
                evidenceId:
                  type: string
                  description: 点亮依据 ID
                score:
                  type: number
                  description: 得分
                remark:
                  type: string
                  description: 备注
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `nodeId` | 是 | string | 知识点节点 ID，对应 knowledge_node.id，兼容 knowledge_node.node_code |

        | Body | `evidenceType` | 否 | string | 点亮依据类型 |

        | Body | `evidenceId` | 否 | string | 点亮依据 ID |

        | Body | `score` | 否 | number | 得分 |

        | Body | `remark` | 否 | string | 备注 |


        主要数据表：knowledge_node、knowledge_node_relation、user_knowledge_progress、user_achievement'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 页码，从 1 开始
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
      description: 每页数量
...
    type:
      name: type
      in: query
      schema:
        type: string
      description: 类型筛选，具体枚举见接口文档数据字典
...
    planId:
      name: planId
      in: path
      required: true
      schema:
        type: string
      description: 学习计划 ID，对应 study_plan.id
...
    nodeId:
      name: nodeId
      in: path
      required: true
      schema:
        type: string
      description: 知识点节点 ID，对应 knowledge_node.id，兼容 knowledge_node.node_code
...
    eventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
      description: 日历事件/学习任务 ID，对应 study_task.id
...
  responses:
    Success:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    PageSuccess:
      description: 分页成功
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/PageResult'
...
  schemas:
    ApiResponse:
      type: object
      required:
      - code
      - message
      - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          nullable: true
          oneOf:
          - type: object
            additionalProperties: true
          - type: array
            items: {}
          - type: string
          - type: number
          - type: boolean
        requestId:
          type: string
          example: req_20260505_xxxx
    PageResult:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          example: 120
        list:
          type: array
          items:
            type: object
            additionalProperties: true